Update run_clm_no_trainer.py
Added a check for the checkpointing step. 
If the checkpointing step is larger than the max training step, it should checkpoint the last step by default.

I made a mistake by putting an extra zero in the checkpointing step which resulted in a checkpointing step larger than my max training step. The mistake was only discovered after I trained it for 3 days and the training has to be done again.